+++ CBSelectCalendarMatrix.m
@@ -18,8 +18,8 @@
    // point where mouseDown occurred in window coordinates
    NSPoint matrixPoint = [self convertPoint:windowPoint fromView:nil];
    // cell where mouseDown occurred
-	int row = 0;
-	int column = 0;
+	NSInteger row = 0;
+	NSInteger column = 0;
 	[self getRow:&row column:&column forPoint:matrixPoint];
 	NSButtonCell * matrixCell = [self cellAtRow:row column:column];
    // if the mouseDown was on a button, perform action
